Crate polars_arrow_rvsry99dx

A native Rust implementation of Apache Arrow, a cross-language development platform for in-memory data.

Currently the project is developed and tested against nightly Rust. To learn more about the status of Arrow in Rust, see README.md.

Modules

array
The central type in Apache Arrow are arrays, represented by the Array trait. An array represents a known-length sequence of values all having the same type.

bitmap
Defines a bitmap, which is used to track which values in an Arrow array are null. This is called a "validity bitmap" in the Arrow documentation.

buffer
The main type in the module is Buffer, a contiguous immutable memory region of fixed size aligned at a 64-byte boundary. MutableBuffer is like Buffer, but it can be mutated and grown.

compute
Computation kernels on Arrow Arrays

csv
Transfer data between the Arrow memory format and CSV (comma-separated values).

datatypes
Defines the logical data types of Arrow arrays.

error
Defines ArrowError for representing failures in various Arrow operations

ipc

json
Transfer data between the Arrow memory format and JSON line-delimited records.

memory
Defines memory-related functions, such as allocate/deallocate/reallocate memory regions, cache and allocation alignments.

record_batch
A two-dimensional batch of column-oriented data with a defined schema.

tensor
Arrow Tensor Type, defined in format/Tensor.fbs.

util