Struct polars_arrow_rvsry99dx::array::PrimitiveArray

pub struct PrimitiveArray<T: ArrowPrimitiveType> { /* fields omitted */ }

Array whose elements are of primitive types.

Implementations

impl<T: ArrowNumericType> PrimitiveArray<T>

Implementation for primitive arrays with numeric types. Boolean arrays are bit-packed and so implemented separately.

pub fn new(
length: usize,
values: Buffer,
null_count: usize,
offset: usize
) -> Self

pub fn values(&self) -> Buffer

Returns a Buffer holding all the values of this array.

Note this doesn't take the offset of this array into account.

pub fn len(&self) -> usize

Returns the length of this array.

pub fn raw_values(&self) -> *const T::Native

Returns a raw pointer to the values of this array.

pub fn value(&self, i: usize) -> T::Native

Returns the primitive value at index i.

Note this doesn't do any bound checking, for performance reason.

pub fn value_slice(&self, offset: usize, len: usize) -> &[T::Native]

Returns a slice for the given offset and length

Note this doesn't do any bound checking, for performance reason.

pub fn builder(capacity: usize) -> PrimitiveBuilder<T>

impl<T: ArrowTemporalType + ArrowNumericType> PrimitiveArray<T> where
i64: From<T::Native>,

pub fn value_as_datetime(&self, i: usize) -> Option<NaiveDateTime>

Returns value as a chrono NaiveDateTime, handling time resolution

If a data type cannot be converted to NaiveDateTime, a None is returned. A valid value is expected, thus the user should first check for validity.

pub fn value_as_date(&self, i: usize) -> Option<NaiveDate>

Returns value as a chrono NaiveDate by using Self::datetime()

If a data type cannot be converted to NaiveDate, a None is returned

pub fn value_as_time(&self, i: usize) -> Option<NaiveTime>

Returns a value as a chrono NaiveTime

Date32 and Date64 return UTC midnight as they do not have time resolution

impl PrimitiveArray<BooleanType>

Specific implementation for Boolean arrays due to bit-packing

pub fn new(
length: usize,
values: Buffer,
null_count: usize,
offset: usize
) -> Self

pub fn values(&self) -> Buffer

Returns a Buffer holds all the values of this array.

Note this doesn't take account into the offset of this array.

pub fn value(&self, i: usize) -> bool

Returns the boolean value at index i.

pub fn builder(capacity: usize) -> BooleanBuilder

impl<T: ArrowTimestampType> PrimitiveArray<T>

pub fn from_vec(data: Vec<i64>, timezone: Option<Arc<String>>) -> Self

Construct a timestamp array from a vec of i64 values and an optional timezone

impl<T: ArrowTimestampType> PrimitiveArray<T>

pub fn from_opt_vec(
data: Vec<Option<i64>>,
timezone: Option<Arc<String>>
) -> Self

Construct a timestamp array from a vec of Option values and an optional timezone