Struct polars_arrow_rvsry99dx::array::StringArray

pub struct StringArray { /* fields omitted */ }

A type of ListArray whose elements are UTF8 strings.

Implementations

impl StringArray

pub fn value(&self, i: usize) -> &str

Returns the element at index i as a string slice.

pub fn value_offset(&self, i: usize) -> i32

Returns the offset for the element at index i.

Note this doesn't do any bound checking, for performance reason.

pub fn value_length(&self, i: usize) -> i32

Returns the length for the element at index i.

Note this doesn't do any bound checking, for performance reason.

pub fn value_offsets(&self) -> Buffer

Returns a clone of the value offset buffer

pub fn value_data(&self) -> Buffer

Returns a clone of the value data buffer

pub fn builder(capacity: usize) -> StringBuilder