Trait polars_arrow_rvsry99dx::array::ArrayBuilder

pub trait ArrayBuilder: Any {
    fn len(&self) -> usize;

    fn append_data(&mut self, data: &[ArrayDataRef]) -> Result<()>;

    fn data_type(&self) -> DataType;

    fn finish(&mut self) -> ArrayRef;

    fn as_any(&self) -> &dyn Any;

    fn as_any_mut(&mut self) -> &mut dyn Any;

    fn into_box_any(self: Box<Self>) -> Box<dyn Any>;
}

Trait for dealing with different array builders at runtime

Required methods

fn len(&self) -> usize

Returns the number of array slots in the builder

fn append_data(&mut self, data: &[ArrayDataRef]) -> Result<()>

Appends data from other arrays into the builder

This is most useful when concatenating arrays of the same type into a builder.

fn data_type(&self) -> DataType

Returns the data type of the builder

This is used for validating array data types in append_data

fn finish(&mut self) -> ArrayRef

Builds the array

fn as_any(&self) -> &dyn Any

Returns the builder as a non-mutable Any reference.

This is most useful when one wants to call non-mutable APIs on a specific builder type. In this case, one can first cast this into a Any, and then use downcast_ref to get a reference on the specific builder.

fn as_any_mut(&mut self) -> &mut dyn Any

Returns the builder as a mutable Any reference.

This is most useful when one wants to call mutable APIs on a specific builder type. In this case, one can first cast this into a Any, and then use downcast_mut to get a reference on the specific builder.

fn into_box_any(self: Box<Self>) -> Box<dyn Any>

Returns the boxed builder as a box of Any.

Loading content...