Struct polars_arrow_rvsry99dx::buffer::Buffer

pub struct Buffer { /* fields omitted */ }

Buffer is a contiguous memory region of fixed size and is aligned at a 64-byte boundary. Buffer is immutable.

Implementations

impl Buffer

pub unsafe fn from_raw_parts(
ptr: *const u8,
len: usize,
capacity: usize
) -> Self

Creates a buffer from an existing memory region (must already be byte-aligned), this Buffer will free this piece of memory when dropped.

Arguments

ptr - Pointer to raw parts

len - Length of raw parts in bytes

capacity - Total allocated memory for the pointer ptr, in bytes

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len bytes.

pub unsafe fn from_unowned(ptr: *const u8, len: usize, capacity: usize) -> Self

Creates a buffer from an existing memory region (must already be byte-aligned), this Buffer does not free this piece of memory when dropped.

Arguments

ptr - Pointer to raw parts

len - Length of raw parts in bytes

capacity - Total allocated memory for the pointer ptr, in bytes

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len bytes.

pub fn len(&self) -> usize

Returns the number of bytes in the buffer

pub fn capacity(&self) -> usize

Returns the capacity of this buffer

pub fn is_empty(&self) -> bool

Returns whether the buffer is empty.

pub fn data(&self) -> &[u8]

Returns the byte slice stored in this buffer

pub fn slice(&self, offset: usize) -> Self

Returns a slice of this buffer, starting from offset.

pub fn raw_data(&self) -> *const u8

Returns a raw pointer for this buffer.

Note that this should be used cautiously, and the returned pointer should not be stored anywhere, to avoid dangling pointers.

pub unsafe fn typed_data<T: ArrowNativeType + Num>(&self) -> &[T]

View buffer as typed slice.

Safety

ArrowNativeType is public so that it can be used as a trait bound for other public components, such as the ToByteSlice trait. However, this means that it can be implemented by user defined types, which it is not intended for.

Also typed_data::<bool> is unsafe as 0x00 and 0x01 are the only valid values for bool in Rust. However, bool arrays in Arrow are bit-packed which breaks this condition.

pub fn empty() -> Self

Returns an empty buffer.