Module polars_arrow_rvsry99dx::compute::kernels::boolean

Defines boolean kernels on Arrow BooleanArray's, e.g. AND, OR and NOT.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

and
Performs AND operation on two arrays. If either left or right value is null then the result is also null.

not
Performs unary NOT operation on an arrays. If value is null then the result is also null.

or
Performs OR operation on two arrays. If either left or right value is null then the result is also null.