Module polars_arrow_rvsry99dx::compute::kernels::comparison

Defines basic comparison kernels for PrimitiveArrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

eq
Perform left == right operation on two arrays.

eq_scalar
Perform left == right operation on an array and a scalar value.

eq_utf8

eq_utf8_scalar

gt
Perform left > right operation on two arrays. Non-null values are greater than null values.

gt_eq
Perform left >= right operation on two arrays. Non-null values are greater than null values.

gt_eq_scalar
Perform left >= right operation on an array and a scalar value. Non-null values are greater than null values.

gt_eq_utf8

gt_eq_utf8_scalar

gt_scalar
Perform left > right operation on an array and a scalar value. Non-null values are greater than null values.

gt_utf8

gt_utf8_scalar

like_utf8

lt
Perform left < right operation on two arrays. Null values are less than non-null values.

lt_eq
Perform left <= right operation on two arrays. Null values are less than non-null values.

lt_eq_scalar
Perform left <= right operation on an array and a scalar value. Null values are less than non-null values.

lt_eq_utf8

lt_eq_utf8_scalar

lt_scalar
Perform left < right operation on an array and a scalar value. Null values are less than non-null values.

lt_utf8

lt_utf8_scalar

neq
Perform left != right operation on two arrays.

neq_scalar
Perform left != right operation on an array and a scalar value.

neq_utf8

neq_utf8_scalar

nlike_utf8

no_simd_compare_op

no_simd_compare_op_scalar