Module polars_arrow_rvsry99dx::ipc::gen::Message

Structs

DictionaryBatch
For sending dictionary encoding information. Any Field can be dictionary-encoded, but in this case none of its children may be dictionary-encoded. There is one vector / column per dictionary, but that vector / column may be spread across multiple dictionary batches by using the isDelta flag

DictionaryBatchArgs

DictionaryBatchBuilder

FieldNode

Message

MessageArgs

MessageBuilder

MessageHeaderUnionTableOffset

RecordBatch
A data header describing the shared memory layout of a "record" or "row" batch. Some systems call this a "row batch" internally and others a "record batch".

RecordBatchArgs

RecordBatchBuilder

Enums

DictionaryBatchOffset

MessageHeader

MessageOffset

RecordBatchOffset

Functions

enum_name_message_header

finish_message_buffer

finish_size_prefixed_message_buffer

get_root_as_message

get_size_prefixed_root_as_message