Struct polars_arrow_rvsry99dx::ipc::reader::FileReader

pub struct FileReader<R: Read + Seek> { /* fields omitted */ }

Arrow File reader

Implementations

impl<R: Read + Seek> FileReader<R>

pub fn try_new(reader: R) -> Result<Self>

Try to create a new file reader

Returns errors if the file does not meet the Arrow Format header and footer requirements

pub fn num_batches(&self) -> usize

Return the number of batches in the file

pub fn schema(&self) -> SchemaRef

Return the schema of the file

pub fn next(&mut self) -> Result<Option<RecordBatch>>

Read the next record batch

pub fn set_index(&mut self, index: usize) -> Result<()>

Read a specific record batch

Sets the current block to the index, allowing random reads