Trait polars_arrow_rvsry99dx::record_batch::RecordBatchReader

pub trait RecordBatchReader {
    fn schema(&mut self) -> SchemaRef;

    fn next_batch(&mut self) -> Result<Option<RecordBatch>>;
}

Definition of record batch reader.

Required methods

fn schema(&mut self) -> SchemaRef

Returns schemas of this record batch reader. Implementation of this trait should guarantee that all record batches returned by this reader should have same schema as returned from this method.

fn next_batch(&mut self) -> Result<Option<RecordBatch>>

Returns next record batch.

Loading content...