Module polars_arrow_rvsry99dx::util::bit_util

Utils for working with bits

Functions

ceil
Returns the ceil of value/divisor

count_set_bits
Returns the number of 1-bits in data

count_set_bits_offset
Returns the number of 1-bits in data, starting from offset with length bits inspected. Note that both offset and length are measured in bits.

get_bit
Returns whether bit at position i in data is set or not

get_bit_raw⚠
Returns whether bit at position i in data is set or not.

round_upto_multiple_of_64
Returns the nearest number that is >= than num and is a multiple of 64

set_bit
Sets bit at position i for data

set_bit_raw⚠
Sets bit at position i for data

set_bits_raw⚠
Sets bits in the non-inclusive range start..end for data